# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Car API",
        "description": "API для управления автомобилями",
        "version": "1.0.0"
    },
    "paths": {
        "/api/cars": {
            "get": {
                "tags": [
                    "Автомобили"
                ],
                "summary": "Получить список всех автомобилей",
                "operationId": "1fd4f1fdc0da258c5084bcdce4bbe693",
                "responses": {
                    "200": {
                        "description": "Список автомобилей",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Car"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Автомобили"
                ],
                "summary": "Создать новый автомобиль",
                "operationId": "6f519ab2c9767d4d3b1be7cf8f1287ed",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Название автомобиля",
                                        "type": "string",
                                        "example": "Toyota Camry"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Автомобиль успешно создан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Car"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректные данные"
                    }
                }
            }
        },
        "/api/cars/{id}": {
            "get": {
                "tags": [
                    "Автомобили"
                ],
                "summary": "Получить информацию об автомобиле по ID",
                "operationId": "b222bfa902eefbb6a4a875762081dcbc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация об автомобиле",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Car"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Автомобили"
                ],
                "summary": "Удалить автомобиль",
                "operationId": "802b64f09e484fbb1bb3817362d30cc9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Автомобиль удален"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Автомобили"
                ],
                "summary": "Обновить информацию об автомобиле",
                "operationId": "067e985e39de9ca94c07247a147d93ca",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Название автомобиля",
                                        "type": "string",
                                        "example": "Toyota Camry"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Автомобиль обновлен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Car"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/configurations": {
            "get": {
                "tags": [
                    "Конфигурации"
                ],
                "summary": "Получить список всех конфигураций",
                "operationId": "ecf557529a783e14266d0c9e51fb28ea",
                "responses": {
                    "200": {
                        "description": "Список конфигураций",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Configuration"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Конфигурации"
                ],
                "summary": "Создать новую конфигурацию",
                "operationId": "a686fda99be004f135bc70ea4b237df3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "car_id",
                                    "name"
                                ],
                                "properties": {
                                    "car_id": {
                                        "description": "ID автомобиля",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "description": "Название конфигурации",
                                        "type": "string",
                                        "example": "Comfort"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Конфигурация успешно создана",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Configuration"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректные данные"
                    }
                }
            }
        },
        "/api/configurations/{id}": {
            "get": {
                "tags": [
                    "Конфигурации"
                ],
                "summary": "Получить одну конфигурацию по ID",
                "operationId": "603f20b4445da5c154c2dd1cb913471a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID конфигурации",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Конфигурация найдена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Configuration"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Конфигурация не найдена"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Конфигурации"
                ],
                "summary": "Удалить конфигурацию",
                "operationId": "c548b7fd6072a3f525f104c851e72e0e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID конфигурации",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Конфигурация удалена"
                    },
                    "404": {
                        "description": "Конфигурация не найдена"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Конфигурации"
                ],
                "summary": "Обновить конфигурацию",
                "operationId": "1d482721a1fd466b52916701c50117ef",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID конфигурации",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "car_id": {
                                        "description": "ID автомобиля",
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "name": {
                                        "description": "Название конфигурации",
                                        "type": "string",
                                        "example": "Premium"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Конфигурация обновлена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Configuration"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Конфигурация не найдена"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Car": {
                "title": "Автомобиль",
                "description": "Модель автомобиля",
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "description": "ID автомобиля",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Название модели автомобиля",
                        "type": "string",
                        "example": "Лада Веста"
                    },
                    "created_at": {
                        "description": "Дата создания записи",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-04-01T10:00:00Z"
                    },
                    "updated_at": {
                        "description": "Дата обновления записи",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-04-01T10:00:00Z"
                    }
                },
                "type": "object"
            },
            "Configuration": {
                "title": "Конфигурация",
                "required": [
                    "car_id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "description": "ID конфигурации",
                        "type": "integer",
                        "example": 1
                    },
                    "car_id": {
                        "description": "ID автомобиля",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Название конфигурации",
                        "type": "string",
                        "example": "Лада Веста"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-04-01T10:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-04-01T10:00:00Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}